User Flags (State, Status, Quest) - user.hasflag, addflag, removeflag - death handling now uses userstate - blink uses userstate - spell chanting uses userstate - panelobject.canuse changed to virtual - clamp now accepts a double, because it's more modular

+++ Chaos/Objects/Monster.cs
@@ -21,7 +21,7 @@ namespace Chaos
         internal DateTime LastUpdate { get; set; }
         internal List<Item> Items { get; set; }
         internal uint Gold;
-        internal override byte HealthPercent => Utility.Clamp<byte>((int)((CurrentHP * 100) / MaximumHP), 0, (int)MaximumHP);
+        internal override byte HealthPercent => Utility.Clamp<byte>((CurrentHP * 100) / MaximumHP, 0, (int)MaximumHP);
         internal override uint MaximumHP { get; }
         internal override uint CurrentHP { get; set; }
